+++ apps/hubble/src/eth/fnameRegistryEventsProvider.test.ts
@@ -5,37 +5,41 @@ import {
 } from './fnameRegistryEventsProvider.js';
 import { jestRocksDB } from '../storage/db/jestUtils.js';
 import Engine from '../storage/engine/index.js';
-import { FarcasterNetwork } from '@farcaster/hub-nodejs';
-import { eip712 } from '@farcaster/core';
+import { ViemLocalEip712Signer, makeUserNameProofClaim, FarcasterNetwork } from '@farcaster/core';
 import { MockHub } from '../test/mocks.js';
 import { getUserNameProof } from '../storage/db/nameRegistryEvent.js';
 import { utf8ToBytes } from '@noble/curves/abstract/utils';
-import { Signer, Wallet } from 'ethers';
+import { generatePrivateKey, privateKeyToAccount, PrivateKeyAccount, Address } from 'viem/accounts';
+import { bytesToHex } from 'viem';
 
 class MockFnameRegistryClient implements FNameRegistryClientInterface {
   private transfersToReturn: FNameTransfer[][] = [];
   private minimumSince = 0;
   private timesToThrow = 0;
-  private signer: Signer;
+  private account: PrivateKeyAccount;
+  private signer: ViemLocalEip712Signer;
+  private signerAddress: Address;
 
-  constructor(signer: Signer) {
-    this.signer = signer;
+  constructor(account: PrivateKeyAccount) {
+    this.account = account;
+    this.signer = new ViemLocalEip712Signer(account);
+    this.signerAddress = account.address;
   }
 
   setTransfersToReturn(transfers: FNameTransfer[][]) {
     this.transfersToReturn = transfers;
     this.transfersToReturn.flat(2).forEach(async (t) => {
       if (t.server_signature === '') {
-        t.server_signature = await this.signer.signTypedData(
-          eip712.EIP_712_USERNAME_DOMAIN,
-          {
-            UserNameProof: eip712.EIP_712_USERNAME_PROOF,
-          },
-          {
-            name: t.username,
-            owner: t.owner,
-            timestamp: t.timestamp,
-          }
+        t.server_signature = bytesToHex(
+          (
+            await this.signer.signUserNameProofClaim(
+              makeUserNameProofClaim({
+                name: t.username,
+                owner: t.owner,
+                timestamp: t.timestamp,
+              })
+            )
+          )._unsafeUnwrap()
         );
       }
     });
@@ -43,6 +47,10 @@ class MockFnameRegistryClient implements FNameRegistryClientInterface {
 
   setMinimumSince(minimumSince: number) {
     this.minimumSince = minimumSince;
+  }
+
+  setSignerAddress(address: Address) {
+    this.signerAddress = address;
   }
 
   throwOnce() {
@@ -63,7 +71,7 @@ class MockFnameRegistryClient implements FNameRegistryClientInterface {
   }
 
   async getSigner(): Promise<string> {
-    return this.signer.getAddress();
+    return this.signerAddress;
   }
 }
 
@@ -73,17 +81,49 @@ describe('fnameRegistryEventsProvider', () => {
   const hub = new MockHub(db, engine);
   let provider: FNameRegistryEventsProvider;
   let mockFnameRegistryClient: MockFnameRegistryClient;
-  const signer = Wallet.createRandom();
+  const account = privateKeyToAccount(generatePrivateKey());
 
   const transferEvents: FNameTransfer[] = [
-    { id: 1, username: 'test1', from: 0, to: 1, timestamp: 1686291736947, owner: signer.address, server_signature: '' },
-    { id: 2, username: 'test2', from: 0, to: 2, timestamp: 1686291740231, owner: signer.address, server_signature: '' },
-    { id: 3, username: 'test3', from: 0, to: 3, timestamp: 1686291751362, owner: signer.address, server_signature: '' },
-    { id: 4, username: 'test3', from: 3, to: 0, timestamp: 1686291752129, owner: signer.address, server_signature: '' },
+    {
+      id: 1,
+      username: 'test1',
+      from: 0,
+      to: 1,
+      timestamp: 1686291736947,
+      owner: account.address,
+      server_signature: '',
+    },
+    {
+      id: 2,
+      username: 'test2',
+      from: 0,
+      to: 2,
+      timestamp: 1686291740231,
+      owner: account.address,
+      server_signature: '',
+    },
+    {
+      id: 3,
+      username: 'test3',
+      from: 0,
+      to: 3,
+      timestamp: 1686291751362,
+      owner: account.address,
+      server_signature: '',
+    },
+    {
+      id: 4,
+      username: 'test3',
+      from: 3,
+      to: 0,
+      timestamp: 1686291752129,
+      owner: account.address,
+      server_signature: '',
+    },
   ];
 
   beforeEach(() => {
-    mockFnameRegistryClient = new MockFnameRegistryClient(signer);
+    mockFnameRegistryClient = new MockFnameRegistryClient(account);
     mockFnameRegistryClient.setTransfersToReturn([
       transferEvents.slice(0, 1),
       transferEvents.slice(1, 2),
@@ -102,15 +142,15 @@ describe('fnameRegistryEventsProvider', () => {
       expect(await getUserNameProof(db, utf8ToBytes('test1'))).toBeTruthy();
       expect(await getUserNameProof(db, utf8ToBytes('test2'))).toBeTruthy();
       await expect(getUserNameProof(db, utf8ToBytes('test4'))).rejects.toThrowError('NotFound');
-      expect((await hub.getHubState())._unsafeUnwrap().lastFnameProof).toEqual(transferEvents[3]!.id);
+      expect((await hub.getHubState())._unsafeUnwrap().lastFnameProof).toEqual(transferEvents[3]?.id);
     });
 
     it('fetches events from where it left off', async () => {
-      await hub.putHubState({ lastFnameProof: transferEvents[0]!.id, lastEthBlock: 0 });
-      mockFnameRegistryClient.setMinimumSince(transferEvents[0]!.id);
+      await hub.putHubState({ lastFnameProof: transferEvents[0]?.id ?? 0, lastEthBlock: 0 });
+      mockFnameRegistryClient.setMinimumSince(transferEvents[0]?.id ?? 0);
       await provider.start();
       expect(await getUserNameProof(db, utf8ToBytes('test2'))).toBeTruthy();
-      expect((await hub.getHubState())._unsafeUnwrap().lastFnameProof).toEqual(transferEvents[3]!.id);
+      expect((await hub.getHubState())._unsafeUnwrap().lastFnameProof).toEqual(transferEvents[3]?.id);
     });
 
     it('does not fail on errors', async () => {
@@ -139,13 +179,33 @@ describe('fnameRegistryEventsProvider', () => {
         from: 0,
         to: 1,
         timestamp: 1686291736947,
-        owner: signer.address,
+        owner: account.address,
         server_signature: '',
       };
       invalidEvent.server_signature = '0x8773442740c17c9d0f0b87022c722f9a136206ed';
       mockFnameRegistryClient.setTransfersToReturn([[invalidEvent]]);
       await provider.start();
       await expect(getUserNameProof(db, utf8ToBytes('test1'))).rejects.toThrowError('NotFound');
+    });
+
+    it('succeeds for a known proof', async () => {
+      const proof = {
+        id: 1,
+        timestamp: 1628882891,
+        username: 'farcaster',
+        owner: '0x8773442740c17c9d0f0b87022c722f9a136206ed' as Address,
+        from: 0,
+        to: 1,
+        user_signature:
+          '0xa6fdd2a69deab5633636f32a30a54b21b27dff123e6481532746eadca18cd84048488a98ca4aaf90f4d29b7e181c4540b360ba0721b928e50ffcd495734ef8471b',
+        server_signature:
+          '0xb7181760f14eda0028e0b647ff15f45235526ced3b4ae07fcce06141b73d32960d3253776e62f761363fb8137087192047763f4af838950a96f3885f3c2289c41b',
+      };
+
+      mockFnameRegistryClient.setTransfersToReturn([[proof]]);
+      mockFnameRegistryClient.setSignerAddress('0xBc5274eFc266311015793d89E9B591fa46294741');
+      await provider.start();
+      expect(await getUserNameProof(db, utf8ToBytes('farcaster'))).toBeTruthy();
     });
   });
 });