+++ packages/core/src/crypto/eip712.ts
@@ -1,17 +1,15 @@
-import { recoverAddress, TypedDataEncoder, Signer, TypedDataDomain, TypedDataField } from 'ethers';
-import { err, Result, ResultAsync } from 'neverthrow';
-import { bytesToHexString, hexStringToBytes } from '../bytes';
-import { HubAsyncResult, HubError, HubResult } from '../errors';
+import { bytesToHex, verifyTypedData } from 'viem';
+import { ResultAsync } from 'neverthrow';
+import { HubAsyncResult, HubError } from '../errors';
 import { VerificationEthAddressClaim } from '../verifications';
-
-export type MinimalEthersSigner = Pick<Signer, 'signTypedData' | 'getAddress'>;
+import { UserNameProofClaim } from '../userNameProof';
 
 export const EIP_712_FARCASTER_DOMAIN = {
   name: 'Farcaster Verify Ethereum Address',
   version: '2.0.0',
   // fixed salt to minimize collisions
-  salt: '0xf2d857f4a3edcb9b78b4d503bfe733db1e3f6cdc2b7971ee739626c97e86a558' as `0x${string}`, // Type cast for viem compatibility
-};
+  salt: '0xf2d857f4a3edcb9b78b4d503bfe733db1e3f6cdc2b7971ee739626c97e86a558',
+} as const;
 
 export const EIP_712_FARCASTER_VERIFICATION_CLAIM = [
   {
@@ -30,105 +28,84 @@ export const EIP_712_FARCASTER_VERIFICATION_CLAIM = [
     name: 'network',
     type: 'uint8',
   },
-];
+] as const;
 
 export const EIP_712_FARCASTER_MESSAGE_DATA = [
   {
     name: 'hash',
     type: 'bytes',
   },
-];
+] as const;
 
 export const EIP_712_USERNAME_DOMAIN = {
   name: 'Farcaster name verification',
   version: '1',
   chainId: 1,
   verifyingContract: '0xe3be01d99baa8db9905b33a3ca391238234b79d1', // name registry contract, will be the farcaster ENS CCIP contract later
-};
+} as const;
 
 export const EIP_712_USERNAME_PROOF = [
   { name: 'name', type: 'string' },
   { name: 'timestamp', type: 'uint256' },
   { name: 'owner', type: 'address' },
-];
+] as const;
 
-export const getSignerKey = async (signer: MinimalEthersSigner): HubAsyncResult<Uint8Array> => {
-  return ResultAsync.fromPromise(signer.getAddress(), (e) => new HubError('unknown', e as Error)).andThen(
-    hexStringToBytes
-  );
-};
-
-export const signVerificationEthAddressClaim = async (
+export const verifyVerificationEthAddressClaimSignature = async (
   claim: VerificationEthAddressClaim,
-  signer: MinimalEthersSigner
-): HubAsyncResult<Uint8Array> => {
-  const hexSignature = await ResultAsync.fromPromise(
-    signer.signTypedData(EIP_712_FARCASTER_DOMAIN, { VerificationClaim: EIP_712_FARCASTER_VERIFICATION_CLAIM }, claim),
-    (e) => new HubError('bad_request.invalid_param', e as Error)
+  signature: Uint8Array,
+  address: Uint8Array
+): HubAsyncResult<boolean> => {
+  const valid = await ResultAsync.fromPromise(
+    verifyTypedData({
+      address: bytesToHex(address),
+      domain: EIP_712_FARCASTER_DOMAIN,
+      types: { VerificationClaim: EIP_712_FARCASTER_VERIFICATION_CLAIM },
+      primaryType: 'VerificationClaim',
+      message: claim,
+      signature,
+    }),
+    (e) => new HubError('unknown', e as Error)
   );
 
-  // Convert hex signature to bytes
-  return hexSignature.andThen((hex) => hexStringToBytes(hex));
+  return valid;
 };
 
-export const verifyVerificationEthAddressClaimSignature = (
-  claim: VerificationEthAddressClaim,
-  signature: Uint8Array
-): HubResult<Uint8Array> => {
-  return recoverSignature(
-    EIP_712_FARCASTER_DOMAIN,
-    { VerificationClaim: EIP_712_FARCASTER_VERIFICATION_CLAIM },
-    claim,
-    signature
-  );
-};
-
-export type UserNameProofClaim = {
-  name: string;
-  timestamp: number;
-  owner: string; // hex address
-};
-
-export const verifyUserNameProof = (nameProof: UserNameProofClaim, signature: Uint8Array): HubResult<Uint8Array> => {
-  return recoverSignature(EIP_712_USERNAME_DOMAIN, { UserNameProof: EIP_712_USERNAME_PROOF }, nameProof, signature);
-};
-
-export const signMessageHash = async (hash: Uint8Array, signer: MinimalEthersSigner): HubAsyncResult<Uint8Array> => {
-  const hexSignature = await ResultAsync.fromPromise(
-    signer.signTypedData(EIP_712_FARCASTER_DOMAIN, { MessageData: EIP_712_FARCASTER_MESSAGE_DATA }, { hash }),
-    (e) => new HubError('bad_request.invalid_param', e as Error)
+export const verifyUserNameProofClaim = async (
+  nameProof: UserNameProofClaim,
+  signature: Uint8Array,
+  address: Uint8Array
+): HubAsyncResult<boolean> => {
+  const valid = await ResultAsync.fromPromise(
+    verifyTypedData({
+      address: bytesToHex(address),
+      domain: EIP_712_USERNAME_DOMAIN,
+      types: { UserNameProof: EIP_712_USERNAME_PROOF },
+      primaryType: 'UserNameProof',
+      message: nameProof,
+      signature,
+    }),
+    (e) => new HubError('unknown', e as Error)
   );
 
-  // Convert hex signature to bytes
-  return hexSignature.andThen((hex) => hexStringToBytes(hex));
+  return valid;
 };
 
-export const verifyMessageHashSignature = (hash: Uint8Array, signature: Uint8Array): HubResult<Uint8Array> => {
-  return recoverSignature(
-    EIP_712_FARCASTER_DOMAIN,
-    { MessageData: EIP_712_FARCASTER_MESSAGE_DATA },
-    { hash },
-    signature
+export const verifyMessageHashSignature = async (
+  hash: Uint8Array,
+  signature: Uint8Array,
+  address: Uint8Array
+): HubAsyncResult<boolean> => {
+  const valid = await ResultAsync.fromPromise(
+    verifyTypedData({
+      address: bytesToHex(address),
+      domain: EIP_712_FARCASTER_DOMAIN,
+      types: { MessageData: EIP_712_FARCASTER_MESSAGE_DATA },
+      primaryType: 'MessageData',
+      message: { hash: bytesToHex(hash) },
+      signature,
+    }),
+    (e) => new HubError('unknown', e as Error)
   );
-};
 
-const recoverSignature = (
-  domain: TypedDataDomain,
-  types: Record<string, Array<TypedDataField>>,
-  claim: Record<string, any>,
-  signature: Uint8Array
-): HubResult<Uint8Array> => {
-  const signatureHexResult = bytesToHexString(signature);
-  if (signatureHexResult.isErr()) {
-    return err(signatureHexResult.error);
-  }
-
-  // Recover address from signature
-  const recoveredHexAddress = Result.fromThrowable(
-    () => recoverAddress(TypedDataEncoder.hash(domain, types, claim), signatureHexResult.value),
-    (e) => new HubError('bad_request.invalid_param', e as Error)
-  )();
-
-  // Convert hex recovered address to bytes
-  return recoveredHexAddress.andThen((hex) => hexStringToBytes(hex));
+  return valid;
 };