+++ packages/core/src/protobufs/typeguards.ts
@@ -208,3 +208,13 @@ export const isMergeStorageAdminRegistryEventHubEvent = (
     typeof event.mergeStorageAdminRegistryEventBody.storageAdminRegistryEvent !== 'undefined'
   );
 };
+
+export const isMergeUsernameProofHubEvent = (
+  event: hubEventProtobufs.HubEvent
+): event is types.MergeUsernameProofHubEvent => {
+  return (
+    event.type === hubEventProtobufs.HubEventType.MERGE_USERNAME_PROOF &&
+    typeof event.mergeUsernameProofBody !== 'undefined' &&
+    typeof event.mergeUsernameProofBody.usernameProof !== 'undefined'
+  );
+};