+++ packages/core/src/signers/eip712Signer.ts
@@ -1,6 +1,7 @@
 import { SignatureScheme } from '../protobufs';
 import { HubAsyncResult } from '../errors';
 import { VerificationEthAddressClaim } from '../verifications';
+import { UserNameProofClaim } from '../userNameProof';
 import { Signer } from './signer';
 
 /**
@@ -16,4 +17,5 @@ export abstract class Eip712Signer implements Signer {
   public abstract getSignerKey(): HubAsyncResult<Uint8Array>;
   public abstract signMessageHash(hash: Uint8Array): HubAsyncResult<Uint8Array>;
   public abstract signVerificationEthAddressClaim(claim: VerificationEthAddressClaim): HubAsyncResult<Uint8Array>;
+  public abstract signUserNameProofClaim(claim: UserNameProofClaim): HubAsyncResult<Uint8Array>;
 }