+++ packages/core/src/signers/ethersV5Eip712Signer.ts
@@ -8,6 +8,7 @@ import { HubAsyncResult, HubError } from '../errors';
 import { eip712 } from '../crypto';
 import { hexStringToBytes } from '../bytes';
 import { VerificationEthAddressClaim } from '../verifications';
+import { UserNameProofClaim } from '../userNameProof';
 
 export type TypedDataSigner = EthersAbstractSigner & EthersTypedDataSigner;
 
@@ -30,7 +31,7 @@ export class EthersV5Eip712Signer extends Eip712Signer {
     const hexSignature = await ResultAsync.fromPromise(
       this._typedDataSigner._signTypedData(
         eip712.EIP_712_FARCASTER_DOMAIN,
-        { MessageData: eip712.EIP_712_FARCASTER_MESSAGE_DATA },
+        { MessageData: [...eip712.EIP_712_FARCASTER_MESSAGE_DATA] },
         { hash }
       ),
       (e) => new HubError('bad_request.invalid_param', e as Error)
@@ -42,8 +43,20 @@ export class EthersV5Eip712Signer extends Eip712Signer {
     const hexSignature = await ResultAsync.fromPromise(
       this._typedDataSigner._signTypedData(
         eip712.EIP_712_FARCASTER_DOMAIN,
-        { VerificationClaim: eip712.EIP_712_FARCASTER_VERIFICATION_CLAIM },
+        { VerificationClaim: [...eip712.EIP_712_FARCASTER_VERIFICATION_CLAIM] },
         claim
+      ),
+      (e) => new HubError('bad_request.invalid_param', e as Error)
+    );
+    return hexSignature.andThen((hex) => hexStringToBytes(hex));
+  }
+
+  public async signUserNameProofClaim(usernameProof: UserNameProofClaim): HubAsyncResult<Uint8Array> {
+    const hexSignature = await ResultAsync.fromPromise(
+      this._typedDataSigner._signTypedData(
+        eip712.EIP_712_USERNAME_DOMAIN,
+        { UserNameProof: [...eip712.EIP_712_USERNAME_PROOF] },
+        usernameProof
       ),
       (e) => new HubError('bad_request.invalid_param', e as Error)
     );