+++ packages/core/src/signers/testUtils.ts
@@ -5,7 +5,9 @@ import { eip712 } from '../crypto';
 import { Factories } from '../factories';
 import { FarcasterNetwork } from '../protobufs';
 import { makeVerificationEthAddressClaim, VerificationEthAddressClaim } from '../verifications';
+import { makeUserNameProofClaim, UserNameProofClaim } from '../userNameProof';
 import { Eip712Signer } from './eip712Signer';
+import { bytesToHex } from 'viem';
 
 export const testEip712Signer = async (signer: Eip712Signer) => {
   let signerKey: Uint8Array;
@@ -20,8 +22,8 @@ export const testEip712Signer = async (signer: Eip712Signer) => {
       const hash = blake3(bytes, { dkLen: 20 });
       const signature = await signer.signMessageHash(hash);
       expect(signature.isOk()).toBeTruthy();
-      const recoveredAddress = eip712.verifyMessageHashSignature(hash, signature._unsafeUnwrap());
-      expect(recoveredAddress).toEqual(ok(signerKey));
+      const valid = await eip712.verifyMessageHashSignature(hash, signature._unsafeUnwrap(), signerKey);
+      expect(valid).toEqual(ok(true));
     });
   });
 
@@ -42,8 +44,8 @@ export const testEip712Signer = async (signer: Eip712Signer) => {
     });
 
     test('succeeds', async () => {
-      const recoveredAddress = eip712.verifyVerificationEthAddressClaimSignature(claim, signature);
-      expect(recoveredAddress).toEqual(ok(signerKey));
+      const valid = await eip712.verifyVerificationEthAddressClaimSignature(claim, signature, signerKey);
+      expect(valid).toEqual(ok(true));
     });
 
     test('succeeds when encoding twice', async () => {
@@ -51,6 +53,52 @@ export const testEip712Signer = async (signer: Eip712Signer) => {
       const signature2 = await signer.signVerificationEthAddressClaim(claim2);
       expect(signature2).toEqual(ok(signature));
       expect(bytesToHexString(signature2._unsafeUnwrap())).toEqual(bytesToHexString(signature));
+    });
+
+    test('fails with HubError', async () => {
+      const result = await signer.signVerificationEthAddressClaim({
+        ...claim,
+        fid: -1n,
+      });
+      expect(result.isErr()).toBe(true);
+      expect(result._unsafeUnwrapErr().errCode).toBe('bad_request.invalid_param');
+    });
+  });
+
+  describe('signUserNameProofClaim', () => {
+    let claim: UserNameProofClaim;
+    let signature: Uint8Array;
+
+    beforeAll(async () => {
+      claim = makeUserNameProofClaim({
+        name: '0x000',
+        timestamp: Date.now(),
+        owner: bytesToHex(signerKey),
+      });
+      const signatureResult = await signer.signUserNameProofClaim(claim);
+      expect(signatureResult.isOk()).toBeTruthy();
+      signature = signatureResult._unsafeUnwrap();
+    });
+
+    test('succeeds', async () => {
+      const valid = await eip712.verifyUserNameProofClaim(claim, signature, signerKey);
+      expect(valid).toEqual(ok(true));
+    });
+
+    test('succeeds when encoding twice', async () => {
+      const claim2: UserNameProofClaim = { ...claim };
+      const signature2 = await signer.signUserNameProofClaim(claim2);
+      expect(signature2).toEqual(ok(signature));
+      expect(bytesToHexString(signature2._unsafeUnwrap())).toEqual(bytesToHexString(signature));
+    });
+
+    test('fails with HubError', async () => {
+      const result = await signer.signUserNameProofClaim({
+        ...claim,
+        timestamp: -1n,
+      });
+      expect(result.isErr()).toBe(true);
+      expect(result._unsafeUnwrapErr().errCode).toBe('bad_request.invalid_param');
     });
   });
 };