+++ packages/core/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @farcaster/core
 
+## 0.10.0
+
+### Minor Changes
+
+- 159c62d: added ViemLocalEip712Signer to exports
+- bfdbfea: Used viem to verify Ethereum signatures
+
+  - Added `signUserNameProofClaim` to `Eip712Signer`
+  - Added `makeUserNameProofClaim`
+
+  - Breaking API changes
+    - make\*Data functions are now async
+    - removed top-level getSignerKey, signVerificationEthAddressClaim, and signMessageHash, use an Eip712Signer class (i.e. EthersEip712Signer or ViemLocalEip712Signer)
+    - rename `verifyUserNameProof` to `verifyUserNameProofClaim`
+
+### Patch Changes
+
+- 159c62d: bumped viem to 1.1.4
+
 ## 0.9.1
 
 ### Patch Changes